Simplify version processing.
In response to the cessation of per-commit tarballs produced by the CI
systems, simplify the version processing code. We did many many many
complicated things to support producing release artifacts that were not
actually supported release artifacts. After a bunch of issues emerged, it was
determined that the best way forward was to stop producing quasi-not-really
non-release tarballs. If people want to consume versions of openstack that
are not released versions, the git repos are always available.

Additionally, this patch removes versioninfo and just uses PKG-INFO and
pkg_resources to handle version processing.

Change-Id: I5c799f3520adb2ba5288d852543706d81a92f8a1